refactor(메뉴그룹): menuGroup 패키지 이동
