+++ src/main/java/kitchenpos/menus/application/MenuService.java
@@ -8,6 +8,7 @@ import kitchenpos.menus.exception.MenuErrorCode;
 import kitchenpos.menus.exception.MenuProductException;
 import kitchenpos.common.domain.ProfanityPolicy;
 import kitchenpos.menus.tobe.domain.menu.*;
+import kitchenpos.menus.tobe.domain.menugroup.MenuGroup;
 import kitchenpos.products.application.ProductService;
 import kitchenpos.products.tobe.domain.Product;
 import org.springframework.stereotype.Service;
