+++ src/main/java/kitchenpos/menus/infra/JpaMenuGroupRepository.java
@@ -1,6 +1,6 @@
 package kitchenpos.menus.infra;
 
-import kitchenpos.menus.tobe.domain.menu.MenuGroup;
+import kitchenpos.menus.tobe.domain.menugroup.MenuGroup;
 import kitchenpos.menus.tobe.domain.menugroup.MenuGroupRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
