+++ src/main/java/kitchenpos/menus/tobe/domain/menu/Menu.java
@@ -4,6 +4,7 @@ import kitchenpos.common.domain.Price;
 import kitchenpos.common.domain.ProfanityPolicy;
 import kitchenpos.menus.exception.MenuErrorCode;
 import kitchenpos.menus.exception.MenuException;
+import kitchenpos.menus.tobe.domain.menugroup.MenuGroup;
 
 import javax.persistence.*;
 import java.math.BigDecimal;
@@ -99,14 +100,6 @@ public class Menu {
         }
     }
 
-    public UUID getId() {
-        return id;
-    }
-
-    public MenuGroup getMenuGroup() {
-        return menuGroup;
-    }
-
     public boolean isDisplayed() {
         return displayed;
     }
@@ -120,6 +113,14 @@ public class Menu {
 
     public void hide() {
         this.displayed = false;
+    }
+
+    public UUID getId() {
+        return id;
+    }
+
+    public MenuGroup getMenuGroup() {
+        return menuGroup;
     }
 
     public MenuProducts getMenuProducts() {
