+++ src/test/java/kitchenpos/menus/application/fixtures/MenuFixture.java
@@ -4,7 +4,7 @@ import kitchenpos.common.domain.ProfanityPolicy;
 import kitchenpos.menus.dto.MenuCreateRequest;
 import kitchenpos.menus.dto.MenuProductRequest;
 import kitchenpos.menus.tobe.domain.menu.Menu;
-import kitchenpos.menus.tobe.domain.menu.MenuGroup;
+import kitchenpos.menus.tobe.domain.menugroup.MenuGroup;
 import kitchenpos.menus.tobe.domain.menu.MenuProduct;
 import kitchenpos.products.tobe.domain.Product;
